src/api/resume_parser.py
import pdfplumber
from docx import Document

def extract_text_from_pdf(file) -> str:
    text = ""
    with pdfplumber.open(file) as pdf:
        for page in pdf.pages:
            page_text = page.extract_text()
            if page_text:
                text += page_text + " "
    return text.strip()

def extract_text_from_docx(file) -> str:
    doc = Document(file)
    return " ".join(p.text for p in doc.paragraphs).strip()
