requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
spacy
nltk
pdfplumber
python-docx
sentence-transformers
matplotlib
seaborn
joblib
torch
seaborn
matplotlib
python-multipart